p00f: rename to p00f, burned result state, standard lint config
Addresses @0xdhrv's review:

- Folder is now extensions/p00f/, manifest name is "p00f", and the
  title is "P00f - Ephemeral Clipboard".
- Adopt the standard store lint toolchain: @raycast/eslint-config,
  eslint, prettier, @types/react, an ESM eslint.config.js (this package
  is type: module), and a fix-lint script. "ray lint" previously
  skipped the formatting check because ESLint was not installed.

Also clears the Greptile notes deferred on the first pass:

- Burn Now switches the result screen to a burned state. It previously
  left the dead link on screen with every action, including a second
  Burn Now, still live.
- Action handlers surface a rejected promise as a Failure toast rather
  than an unhandled rejection.
- The two no-view commands split the toast title and message instead of
  putting raw error text in the title.

Verified with a clean npm ci plus ray build and ray lint at the new
path, and an 18-check create/info/reveal/burn round-trip against the
live service covering text, PNG bytes, reveal budget exhaustion,
owner-token burn, PIN rejection, and reveal-anchored arming.
