+++ extensions/p00f/CHANGELOG.md
@@ -7,5 +7,5 @@
 - **Poof Clipboard**: quick command for one clipboard file, plain text, HTML as text fallback, or a clipboard image (such as a `Cmd+Shift+Ctrl+4` screenshot).
 - All Links and owner tokens copied with Raycast's concealed clipboard option.
 - Optional paste-after-create preference.
-- Result screen with copy link, paste link, copy owner token, burn now, and open in browser actions.
+- Result screen with copy link, paste link, copy owner token, burn now, and open in browser actions. Burning switches the screen to a burned state so the dead link is never presented as live.
 - Anonymous machine-path client with no client-identifying headers. The hosted p00f service receives only ciphertext.
