+++ extensions/p00f/eslint.config.js
@@ -1,0 +1,4 @@
+import { defineConfig } from "eslint/config";
+import raycast from "@raycast/eslint-config";
+
+export default defineConfig([...raycast]);
