+++ extensions/p00f/package.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://www.raycast.com/schemas/extension.json",
-  "name": "p00f-raycast",
-  "title": "P00f",
+  "name": "p00f",
+  "title": "P00f - Ephemeral Clipboard",
   "description": "Create zero-knowledge, ephemeral p00f links from Raycast.",
   "type": "module",
   "icon": "icon.png",
@@ -173,6 +173,7 @@
     "dev": "ray develop",
     "build": "ray build -e dist",
     "lint": "ray lint",
+    "fix-lint": "ray lint --fix",
     "check": "npm run build && npm run lint",
     "publish": "npx @raycast/api@latest publish"
   },
@@ -181,7 +182,11 @@
     "@raycast/api": "^1.104.20"
   },
   "devDependencies": {
+    "@raycast/eslint-config": "^2.2.0",
     "@types/node": "^26.0.0",
+    "@types/react": "^19.2.17",
+    "eslint": "^10.8.0",
+    "prettier": "^3.9.6",
     "typescript": "^6.0.3"
   }
 }
