+++ extensions/p00f/src/lib/clipboard.ts
